refactor: consolidate tests and add dependency injection
- Merge *_fixes.py files into main test files to reduce clutter
- Add session-scoped fixtures in conftest.py for expensive objects
- Implement dependency injection for align() function with pipeline caching
- Add pytest.mark.parametrize for CLI format tests
- Add clear_pipeline_cache() function for testing and memory management
- Create dependency injection demo script
- Fix all import issues and ensure 212 tests pass

Files consolidated:
- test_api_fixes.py → test_api.py
- test_segmenter_fixes.py → test_segmenter.py
- test_parser_fixes.py → test_parser.py
- test_aligner_fixes.py → test_align.py

Performance improvements:
- Pipeline caching reduces initialization overhead
- Dependency injection allows custom pipeline reuse
- Session-scoped fixtures prevent redundant object creation

Test improvements:
- Parameterized CLI tests for better reporting
- Cleaner test organization and maintainability
- All tests passing with improved structure
