--- tests/test_parser_fixes.py
@@ -1,139 +1,0 @@
-"""
-Tests for the parser fixes to ensure stateless behavior and improved filtering.
-"""
-
-from pathlib import Path
-
-import pytest
-
-from vivre.parser import VivreParser
-
-
-class TestParserStatelessness:
-    """Test that the parser is truly stateless and can be reused safely."""
-
-    def test_parser_reuse_no_state_pollution(self):
-        """Test that reusing a parser doesn't cause state pollution."""
-        parser = VivreParser()
-
-        # Find test files
-        test_files = list(Path("tests/data").glob("*.epub"))
-        if len(test_files) < 2:
-            pytest.skip("Need at least 2 test EPUB files")
-
-        file1, file2 = test_files[0], test_files[1]
-
-        # Parse first file
-        chapters1 = parser.parse_epub(file1)
-
-        # Parse second file - should not be affected by first file's metadata
-        chapters2 = parser.parse_epub(file2)
-
-        # Both should work independently
-        assert isinstance(chapters1, list)
-        assert isinstance(chapters2, list)
-
-        # The parser should not retain state between calls
-        # This test verifies that the parser is truly stateless
-
-
-class TestImprovedFiltering:
-    """Test the improved non-story content filtering."""
-
-    def test_href_priority_over_title(self):
-        """Test that href patterns are prioritized over title patterns."""
-        parser = VivreParser()
-
-        # Test case 1: File with non-story href but story-like title
-        # This should be filtered out (href takes priority)
-        result1 = parser._is_non_story_content(
-            title="Chapter 1: The Beginning", href="copyright.xhtml", book_language="en"
-        )
-        assert (
-            result1 is True
-        ), "copyright.xhtml should be filtered despite story-like title"
-
-        # Test case 2: File with story-like href but non-story title
-        # This should NOT be filtered out (href takes priority)
-        result2 = parser._is_non_story_content(
-            title="Important Notice", href="chapter01.xhtml", book_language="en"
-        )
-        assert (
-            result2 is False
-        ), "chapter01.xhtml should not be filtered despite non-story title"
-
-        # Test case 3: File with non-story href and non-story title
-        # This should be filtered out
-        result3 = parser._is_non_story_content(
-            title="Table of Contents", href="toc.xhtml", book_language="en"
-        )
-        assert result3 is True, "toc.xhtml should be filtered"
-
-        # Test case 4: File with story-like href and story-like title
-        # This should NOT be filtered out
-        result4 = parser._is_non_story_content(
-            title="Chapter 1: The Adventure Begins",
-            href="chapter01.xhtml",
-            book_language="en",
-        )
-        assert result4 is False, "chapter01.xhtml should not be filtered"
-
-
-class TestXMLNamespaces:
-    """Test that XML namespaces are handled correctly."""
-
-    def test_metadata_extraction_with_namespaces(self):
-        """Test that metadata extraction works with proper namespaces."""
-        parser = VivreParser()
-
-        # This test would require a real EPUB file with proper namespaces
-        # For now, we'll test that the method exists and has the right signature
-        assert hasattr(parser, "_extract_metadata")
-        # Check that the return type annotation exists and is correct
-        import typing
-
-        assert (
-            parser._extract_metadata.__annotations__["return"] == typing.Dict[str, str]
-        )
-
-
-class TestParserRobustness:
-    """Test that the parser is more robust with the fixes."""
-
-    def test_parser_handles_multiple_files(self):
-        """Test that the parser can handle multiple files without issues."""
-        parser = VivreParser()
-
-        # Find test files
-        test_files = list(Path("tests/data").glob("*.epub"))
-        if len(test_files) < 2:
-            pytest.skip("Need at least 2 test EPUB files")
-
-        results = []
-        for epub_file in test_files[:2]:  # Test with first 2 files
-            try:
-                chapters = parser.parse_epub(epub_file)
-                results.append((epub_file, len(chapters), True))
-            except Exception:  # noqa: BLE001
-                results.append((epub_file, 0, False))
-
-        # All files should be parsed successfully
-        for file_path, chapter_count, success in results:
-            assert success, f"Failed to parse {file_path}"
-            assert chapter_count >= 0, f"Invalid chapter count for {file_path}"
-
-
-def test_parser_documentation():
-    """Test that the parser documentation reflects the stateless nature."""
-    parser = VivreParser()
-
-    # Check that the docstring mentions stateless behavior
-    doc = parser.__doc__
-    assert doc is not None
-    assert "stateless" in doc.lower()
-    assert "reuse" in doc.lower()
-
-    # Check that the parse_epub method doesn't mention instance variables
-    parse_doc = parser.parse_epub.__doc__
-    assert parse_doc is not None
-    # Should not mention _book_title or _book_language in the context of instance variables
